Document/requirements.txt.txt
# Core Backend
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.3
python-dotenv==1.0.1

# Frontend
streamlit==1.35.0

# Machine Learning & Data Processing
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4

# IBM Watsonx (Granite LLM Integration)
ibm-watson-machine-learning==1.0.357  # or latest available

# Pinecone for semantic search
pinecone-client==3.2.2

# Utility Libraries
requests==2.32.2
aiofiles==23.2.1  # For async file uploads
python-multipart==0.0.9  # For form uploads in FastAPI

# Optional (Visualization, Plotting)
matplotlib==3.9.0
seaborn==0.13.2
